Extract testhelpers + added test file for courselists
